Split out common egmde::PrimarySelectionDeviceController from the two primary selection protocols.
